move Node to field component for simpler submit handling and fixing its label
